+++ src/main/java/com/drathonix/experiencedworlds/unified/CLIENTRY.java
@@ -2,7 +2,7 @@ package com.drathonix.experiencedworlds.unified;
 
 
 //? if fabric {
-/*import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.api.ClientModInitializer;
 
 public class CLIENTRY implements ClientModInitializer {
         @Override
@@ -10,7 +10,7 @@ public class CLIENTRY implements ClientModInitializer {
 
         }
 }
-*///?}
+//?}
 
 //? elif forge {
 /*import com.vicious.loadmychunks.forge.LMCForge;
@@ -29,7 +29,7 @@ public class LMCCLIENTRY {
 *///?}
 
 //? elif neoforge {
-import com.drathonix.experiencedworlds.ExperiencedWorlds;
+/*import com.drathonix.experiencedworlds.ExperiencedWorlds;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 //? if >=1.20.6
@@ -37,7 +37,7 @@ import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 //? if <1.20.6
-/*@Mod.EventBusSubscriber(modid=LoadMyChunks.MOD_ID,bus= Mod.EventBusSubscriber.Bus.MOD,value= Dist.CLIENT)*/
+/^@Mod.EventBusSubscriber(modid=LoadMyChunks.MOD_ID,bus= Mod.EventBusSubscriber.Bus.MOD,value= Dist.CLIENT)^/
 //? if >=1.20.6
 @EventBusSubscriber(modid= ExperiencedWorlds.MOD_ID, bus= EventBusSubscriber.Bus.MOD, value= Dist.CLIENT)
 public class CLIENTRY {
@@ -45,5 +45,5 @@ public class CLIENTRY {
     public static void clientInit(FMLClientSetupEvent event) {
     }
 }
-//?}
+*///?}