+++ src/main/java/com/drathonix/serverstatistics/common/bridge/IMixinDimensionDataStorage.java
@@ -5,4 +5,6 @@ import java.io.File;
 public interface IMixinDimensionDataStorage {
     void ss$forceSave(String key);
     File ss$getCustomDataFile(String name);
+
+    File ss$getWorldDir();
 }