+++ src/main/java/com/drathonix/serverstatistics/common/event/AdvancedFirstTimeEvent.java
@@ -2,6 +2,7 @@ package com.drathonix.serverstatistics.common.event;
 
 import net.minecraft.advancements.Advancement;
 import net.minecraft.server.level.ServerPlayer;
+import org.jetbrains.annotations.Nullable;
 
 public class AdvancedFirstTimeEvent  {
     private final ServerPlayer player;
@@ -10,7 +11,7 @@ public class AdvancedFirstTimeEvent  {
         this.player=sp;
         this.advancementKey=key;
     }
-    public ServerPlayer getPlayer(){
+    public @Nullable ServerPlayer getPlayer(){
         return player;
     }
     public Advancement getAdvancement(){